check if full profile.db in v30_to_v31.py

+++ anvio/migrations/profile/v30_to_v31.py
@@ -27,10 +27,14 @@ def migrate(db_path):
     if str(profile_db.get_version()) != current_version:
         raise ConfigError("Version of this profile database is not %s (hence, this script cannot really do anything)." % current_version)
 
+    tables_in_db = profile_db.get_table_names()
+    is_full_profile = 'mean_coverage_Q2Q3_splits' in tables_in_db or 'atomic_data_splits' in tables_in_db
 
-    # contigs_ordered -> items_ordered
-    profile_db.set_meta_value('items_ordered', profile_db.get_meta_value('contigs_ordered'))
-    profile_db.remove_meta_key_value_pair('contigs_ordered')
+    if is_full_profile:
+        # contigs_ordered -> items_ordered
+        profile_db.set_meta_value('items_ordered', profile_db.get_meta_value('contigs_ordered'))
+        profile_db.remove_meta_key_value_pair('contigs_ordered')
+
 
     # set the version
     profile_db.remove_meta_key_value_pair('version')
@@ -40,8 +44,12 @@ def migrate(db_path):
     profile_db.disconnect()
     progress.end()
 
-    run.info_single("Your profile db is now %s (WRONG .. anvi'o never takes breaks)." % next_version, nl_after=1, nl_before=1, mc='green')
-
+    if is_full_profile:
+        run.info_single("Your profile db is now %s (WRONG .. anvi'o never takes breaks)." % next_version, nl_after=1, nl_before=1, mc='green')
+    else:
+        run.info_single("Your profile db is now version %s. But essentially nothing really happened to your\
+                         database since it was a blank profile (which is OK, move along)." \
+                                                            % next_version, nl_after=1, nl_before=1, mc='green')
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='A simple script to upgrade profile database from version %s to version %s' % (current_version, next_version))
